♻️ 41676W task: decompose sqlite cache keys
