+++ .agentplane/tasks/202605290508-41676W/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202605290508-41676W/sqlite-task-cache-decomposition",
+  "created_at": "2026-05-29T05:08:26.986Z",
+  "diffstat_sha256": "sha256:dd6e619bda6e470d15bb7413d64ff80d615cab19bbb73ed6c5e65f073c3add5e",
+  "last_verified_at": "2026-05-29T05:13:51.058Z",
+  "last_verified_diffstat_sha256": "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
+  "schema_version": 1,
+  "task_id": "202605290508-41676W",
+  "updated_at": "2026-05-29T05:08:26.986Z",
+  "verify": {
+    "status": "pass"
+  }
+}
