+++ .agentplane/tasks/202605290508-41676W/pr/review.md
@@ -1,0 +1,39 @@
+# PR Review
+
+Created: 2026-05-29T05:08:26.986Z
+
+## Task
+
+- Task: `202605290508-41676W`
+- Title: SQLite task cache decomposition
+- Status: DOING
+- Branch: `task/202605290508-41676W/sqlite-task-cache-decomposition`
+- Canonical task record: `.agentplane/tasks/202605290508-41676W/README.md`
+
+## Verification
+
+- State: ok
+- Note: SQLite task projection cache helpers extracted into local-task-sqlite-cache-key.ts; local-task-sqlite-cache.ts reduced from hotspot range to 236 lines. Also resolved a main-branch knip blocker in the Obsidian projection facade by making the exported file type a real facade import/re-export.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T05:08:26.986Z
+- Branch: task/202605290508-41676W/sqlite-task-cache-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+ .../task-backend/local-task-sqlite-cache-key.ts    | 212 ++++++++++++++++++++
+ .../task-backend/local-task-sqlite-cache.ts        | 222 +--------------------
+ packages/agentplane/src/commands/task/obsidian.ts  |   3 +-
+ 3 files changed, 225 insertions(+), 212 deletions(-)
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
